+++ webapp/src/app/projects/[projectName]/[languageName]/page.tsx
@@ -1,10 +1,21 @@
 'use client';
 
 import { NextPage } from 'next';
-import React, { useCallback, useEffect, useState } from 'react';
+import { useSearchParams } from 'next/navigation';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import ErrorIcon from '@mui/icons-material/Error';
-import { Box, CircularProgress, Typography } from '@mui/material';
+import ListIcon from '@mui/icons-material/List';
+import SearchIcon from '@mui/icons-material/Search';
+import {
+  Box,
+  CircularProgress,
+  Tab,
+  TextField,
+  Typography,
+} from '@mui/material';
+import { TabContext, TabList, TabPanel } from '@mui/lab';
 
+import Breadcrumbs from '@/components/Breadcrumbs';
 import Sidebar from '@/components/Sidebar';
 import TitleBar from '@/components/TitleBar';
 import MessageList from '@/components/MessageList';
@@ -13,6 +24,8 @@ import SidebarContextProvider from '@/components/SidebarContext';
 import Header from '@/components/Header';
 import Main from '@/components/Main';
 import { LanguageResponse } from '@/app/api/projects/[projectName]/languages/[languageId]/messages/route';
+import SearchContextProvider, { SearchState } from '@/components/SearchContext';
+import { textIncludesQuery } from '@/utils/search';
 
 type LanguageLoadingState = {
   language: undefined;
@@ -34,16 +47,32 @@ type LanguageState =
   | LanguageNotFoundState
   | LanguageReadyState;
 
+type TabId = 'tree' | 'find';
+
 const MessagesPage: NextPage<{
   params: { languageName: string; messageId?: string; projectName: string };
 }> = ({ params }) => {
   const { languageName, projectName } = params;
 
   const [messageId, setMessageId] = useState(params.messageId);
-  const [state, setLanguageState] = useState<LanguageState>({
+  const [languageState, setLanguageState] = useState<LanguageState>({
     language: undefined,
     status: 'loading',
   });
+  const searchParams = useSearchParams();
+  const initialQuery = searchParams.get('query');
+  const [tab, setTab] = useState<TabId>(initialQuery ? 'find' : 'tree');
+  const [searchState, setSearchState] = useState<SearchState>(
+    initialQuery
+      ? {
+          query: initialQuery,
+          status: 'busy',
+        }
+      : {
+          query: '',
+          status: 'idle',
+        },
+  );
 
   useEffect(() => {
     fetch(`/api/projects/${projectName}/languages/${languageName}/messages`)
@@ -63,6 +92,16 @@ const MessagesPage: NextPage<{
       });
   }, [projectName, languageName]);
 
+  const onChangeTab = useCallback((_e: React.SyntheticEvent, newTab: TabId) => {
+    if (newTab !== 'find') {
+      window.history.pushState(null, '', window.location.pathname);
+    }
+    if (newTab === 'find') {
+      setSearchState({ query: '', status: 'idle' });
+    }
+    setTab(newTab);
+  }, []);
+
   const onItemSelectionToggle = useCallback(
     (_e: React.SyntheticEvent, id: string, isSelected: boolean) => {
       if (isSelected) {
@@ -77,17 +116,171 @@ const MessagesPage: NextPage<{
     [languageName, projectName],
   );
 
+  const onChangeSearchQuery = useCallback(
+    (event: React.ChangeEvent<HTMLInputElement>) => {
+      if (event.target.value === '') {
+        setSearchState({
+          query: '',
+          status: 'idle',
+        });
+        window.history.pushState(null, '', window.location.pathname);
+      } else {
+        setSearchState({
+          query: event.target.value,
+          status: 'busy',
+        });
+        const params = new URLSearchParams();
+        params.set('query', event.target.value);
+        const path = ['projects', projectName, languageName];
+        if (messageId) {
+          path.push(messageId);
+        }
+        window.history.pushState(null, '', `/${path.join('/')}?${params}`);
+      }
+    },
+    [languageName, messageId, projectName],
+  );
+
+  const messages = useMemo(() => {
+    if (languageState.status !== 'ready') {
+      return [];
+    }
+
+    if (tab === 'tree' || (tab === 'find' && searchState.status === 'idle')) {
+      if (messageId) {
+        return languageState.language.messages.filter((m) =>
+          m.id.startsWith(messageId),
+        );
+      }
+      return languageState.language.messages;
+    }
+
+    if (tab === 'find') {
+      return languageState.language.messages.filter((m) => {
+        const translation = languageState.language.translations[m.id];
+        return (
+          textIncludesQuery(m.id, searchState.query) ||
+          textIncludesQuery(m.defaultMessage, searchState.query) ||
+          (translation &&
+            textIncludesQuery(translation.text, searchState.query))
+        );
+      });
+    }
+
+    return [];
+  }, [
+    languageState.language?.messages,
+    languageState.language?.translations,
+    languageState.status,
+    messageId,
+    searchState.query,
+    searchState.status,
+    tab,
+  ]);
+
   return (
-    <Box sx={{ display: 'flex', minHeight: '100dvh' }}>
-      <SidebarContextProvider>
-        <Header
-          languageName={languageName}
-          messageId={messageId}
-          projectName={projectName}
-        />
-        <Sidebar>
-          <TitleBar languageName={languageName} projectName={projectName} />
-          {state.status === 'loading' && (
+    <SearchContextProvider value={searchState}>
+      <Box sx={{ display: 'flex', minHeight: '100dvh' }}>
+        <SidebarContextProvider>
+          <Header>
+            {tab === 'find' && searchState.status === 'busy' ? (
+              <Box>
+                <Typography component="span">Search results for</Typography>{' '}
+                <Typography
+                  component="span"
+                  sx={{ backgroundColor: 'yellow', p: 1 }}
+                >
+                  {searchState.query}
+                </Typography>
+              </Box>
+            ) : (
+              <Breadcrumbs
+                languageName={languageName}
+                messageId={messageId}
+                projectName={projectName}
+              />
+            )}
+          </Header>
+          <Sidebar>
+            <TitleBar languageName={languageName} projectName={projectName} />
+            <TabContext value={tab}>
+              <Box
+                sx={{
+                  borderBottom: 1,
+                  borderColor: 'divider',
+                }}
+              >
+                <TabList
+                  aria-label="Navigation"
+                  onChange={onChangeTab}
+                  sx={{
+                    '& .MuiTab-root': {
+                      minHeight: 48,
+                    },
+                    minHeight: 48,
+                  }}
+                  variant="fullWidth"
+                >
+                  <Tab
+                    icon={<ListIcon />}
+                    iconPosition="start"
+                    label="List"
+                    value="tree"
+                  />
+                  <Tab
+                    icon={<SearchIcon />}
+                    iconPosition="start"
+                    label="Find"
+                    value="find"
+                  />
+                </TabList>
+              </Box>
+              <TabPanel sx={{ overflowX: 'auto', padding: 0 }} value="tree">
+                {languageState.status === 'loading' && (
+                  <Box
+                    alignItems="center"
+                    display="flex"
+                    flexDirection="column"
+                    height="100%"
+                    justifyContent="center"
+                  >
+                    <CircularProgress />
+                  </Box>
+                )}
+                {languageState.status === 'ready' && (
+                  <MessageTree
+                    messageId={messageId}
+                    messages={languageState.language.messages}
+                    onItemSelectionToggle={onItemSelectionToggle}
+                  />
+                )}
+              </TabPanel>
+              <TabPanel
+                sx={{
+                  padding: 0,
+                }}
+                value="find"
+              >
+                <Box height="100%" p={2}>
+                  <TextField
+                    label="Search"
+                    onChange={onChangeSearchQuery}
+                    value={searchState.query}
+                    variant="outlined"
+                  />
+                  <Box>
+                    <Typography variant="caption">
+                      {searchState.status === 'busy' &&
+                        `${messages.length} results`}
+                    </Typography>
+                  </Box>
+                </Box>
+              </TabPanel>
+            </TabContext>
+          </Sidebar>
+        </SidebarContextProvider>
+        <Main>
+          {languageState.status === 'loading' && (
             <Box
               alignItems="center"
               display="flex"
@@ -98,57 +291,50 @@ const MessagesPage: NextPage<{
               <CircularProgress />
             </Box>
           )}
-          {state.status === 'ready' && (
-            <MessageTree
-              messageId={messageId}
-              messages={state.language.messages}
-              onItemSelectionToggle={onItemSelectionToggle}
-            />
+          {languageState.status === 'not-found' && (
+            <Box
+              alignItems="center"
+              display="flex"
+              flexDirection="column"
+              height="100%"
+              justifyContent="center"
+            >
+              <ErrorIcon />
+              <Typography component="h1" fontWeight="bold">
+                Not Found
+              </Typography>
+            </Box>
           )}
-        </Sidebar>
-      </SidebarContextProvider>
-      <Main>
-        {state.status === 'loading' && (
-          <Box
-            alignItems="center"
-            display="flex"
-            flexDirection="column"
-            height="100%"
-            justifyContent="center"
-          >
-            <CircularProgress />
-          </Box>
-        )}
-        {state.status === 'not-found' && (
-          <Box
-            alignItems="center"
-            display="flex"
-            flexDirection="column"
-            height="100%"
-            justifyContent="center"
-          >
-            <ErrorIcon />
-            <Typography component="h1" fontWeight="bold">
-              Not Found
-            </Typography>
-          </Box>
-        )}
-        {state.status === 'ready' && (
-          <MessageList
-            languageName={languageName}
-            messages={
-              messageId
-                ? state.language.messages.filter((m) =>
-                    m.id.startsWith(messageId),
-                  )
-                : state.language.messages
-            }
-            projectName={projectName}
-            translations={state.language.translations}
-          />
-        )}
-      </Main>
-    </Box>
+          {languageState.status === 'ready' && (
+            <>
+              {tab === 'find' &&
+              searchState.status === 'busy' &&
+              messages.length === 0 ? (
+                <Box
+                  alignItems="center"
+                  display="flex"
+                  flexDirection="column"
+                  height="100%"
+                  justifyContent="center"
+                >
+                  <ErrorIcon fontSize="large" />
+                  <Typography component="span" fontWeight="bold" variant="h6">
+                    No messages found matching &apos;{searchState.query}&apos;
+                  </Typography>
+                </Box>
+              ) : (
+                <MessageList
+                  languageName={languageName}
+                  messages={messages}
+                  projectName={projectName}
+                  translations={languageState.language.translations}
+                />
+              )}
+            </>
+          )}
+        </Main>
+      </Box>
+    </SearchContextProvider>
   );
 };